New image label definition: "register_default_instance"
A new module instance can ask to be registered as "first instance of module"
at cluster and node level. This information is useful to discover other
modules and establish default behaviours when further module instances are installed.

Modules providing services to other modules are the target of this new feature.