fix(allocate-ports): avoid startup timeout
The "cluster.userdomains" library relies on the Redis SCAN operation, which
may be too slow when executed on a worker node due to network
latency. In some rare cases, obtaining the domain list takes longer than
90 seconds, and ldapproxy.service fails to start.

Connect to the local Redis replica to avoid network latency and increase
the SCAN speed by 100 times.